requirements.txt
Flask==3.1.2
flask-cors==4.0.0
google-generativeai==0.8.5
pinecone==7.3.0
python-dotenv==1.2.1
requests==2.31.0
